fixed anchor paths  and navigate options
If the anchor path was not specified and the path to walk was
absolute, the cwd was prepended. This does not make sense. Now, if the
anchor path is NOT specified and the path to walk is absolute, the cwd
is NOT prepended. The cwd is only prepended if the path to walk is
relative, i.e. no / as first char of path

navigate options would cause a case clause exception. An invalid
option now causes handlebar to abort with a sane message